fix: avoid FOUC when dark mode is persisted in local storage
This commit fixes an issue when a user's preference is dark mode.
Before, when the page is reloaded, a 'flicker' of the light mode would
appear as it was not guaranteed that their theme preference was the
first action conducted within the browser. This is commonly known as a
Flash Of Unstyled Content (FOUC).

A description of this issue can be found
[here](https://stackoverflow.com/questions/3221561/eliminate-flash-of-unstyled-content).

This commit ensures that the initialisation of a users preference is the
first action done by the browser, which stops the flickering effect.