feat: refactor `modal.service` to singleton service
This update allows the `modal.service` to create multiple modal
components with unique IDs which can be called throughout the Angular
project. Each modal can be independently toggled from one another. The
modal service uses RxJs to update each modal components' state.

There is still some work to be done on the modal service and component.
For example, adding an initial positional argument for each modal to that
it can be positioned in the view.

A future milestone for this feature is to dynamically load/reload modal
components into the DOM using `ngOnDestroy` or Angulars' `Renderer2`
module. This would ensure there are no memory leaks and would provide
the best basis for a personal Angular module for future projects.