feat: automatically focus command palette when opening
An auto-focus directive has been made so that items can be automatically
focused once rendered in the DOM by adding the `[focus]=true` attribute
to a HTML attribute.

Currently, this works in `modules.command-palette.component` by using an
`ngIf` conditional on the input tag. In the future, I hope to add to
satisfy the DOM manipulation inside the `shared.modal.component` instead
of having to rely on an ambiguous DOM manipulation within the HTML.