feat: add fuzzy-search interface to `command-palette.component`
This commit uses [Fuse.js](https://fusejs.io) as a fuzzy-finder to
search for the plugin name and plugin company, and returns the most
likely matches from a user's input.

At the moment, a sample JSON file of plugins is used as a service which
interacts with the Fuse.js `search()` functionality.