feat: add dynamic search functionality to `command-palette.component`
This commit adds basic conditional matching on the input search tag in
`command-palette.component` so it can be used to conditionally render
HTML elements.

At the moment, when the search box is empty, a placeholder for user
history and user favourites is shown (this will be developed in later
commits). When the user starts typing into the input box, HTML is
rendered showing the `resultsFound` HTML template.