feat: implement logic for `services.favourites.service`
This commits adds an option for users to favourite certain plugins -
these plugins are subsequently persisted in the users local storage so
they can be recalled on page reload.

There are also some edge cases for the history and favourites
interaction which have been handled. For example, if a user has added
plugin `a` to their favourites, it will not be shown in the users search
history, hence preventing duplicate entries in the
`command-palette-body.component` template.